feat: reorganize source files
