+++ mod.ts
@@ -1,350 +1,33 @@
 /**
  * Steno: A fast Deno-powered static site generator.
  *
- * This module exports the main {@link Steno} orchestrator class, the {@link Theme} class for managing theme templates
- * and assets, and Scribe rendering utilities.
- *
- * @example
- * ```ts
- * import { Steno } from "@steno/steno";
- *
- * const steno = new Steno("content/.steno/config.yml");
- * await steno.build();
- * ```
- *
  * @module
  */
 
-import { loadConfig, loadPlugins, type SiteConfig } from "./src/config.ts";
-import type { StenoPlugin } from "./src/plugins.ts";
-import { Theme } from "./src/theme/theme.ts";
-import type { StenoTheme } from "./src/theme/types.ts";
-import { ensureDirSync } from "./src/fileUtils.ts";
-import { parseFrontmatter } from "./src/frontmatter.ts";
-import { startDevServer } from "./src/server.ts";
-import { parseCliArgs, printHelp } from "./src/cli.ts";
-import { runAstTransforms, runHtmlTransforms } from "./src/plugins.ts";
-import { buildCollections, type CollectionMap } from "./src/collections.ts";
+import { printHelp } from "./src/utils/cli.ts";
+import { runStenoCli } from "./src/core/steno.ts";
+
+export { Steno, runStenoCli, type StenoHooks } from "./src/core/steno.ts";
 export type {
   Collection,
   CollectionItem,
   CollectionMap,
-} from "./src/collections.ts";
-import { marked } from "marked";
-import { dirname, isAbsolute, join } from "@std/path";
-
-export { filters, render } from "./src/scribe.ts";
-export type { ScribeOptions } from "./src/scribe.ts";
-export type { SiteConfig } from "./src/config.ts";
-export type { StenoTheme } from "./src/theme/types.ts";
+} from "./src/core/collections.ts";
+export { filters, render } from "./src/utils/scribe.ts";
+export type { ScribeOptions } from "./src/utils/scribe.ts";
+export type {
+  CollectionConfig,
+  PluginEntry,
+  SiteConfig,
+  StenoPlugin,
+  StenoTheme,
+  ThemeConfigField,
+} from "./src/types.ts";
 export { Theme } from "./src/theme/theme.ts";
-export type { StenoPlugin } from "./src/plugins.ts";
-
-export interface StenoHooks {
-  /**
-   * Called before the build process starts.
-   * @param config The site configuration.
-   */
-  beforeBuild?: (config: SiteConfig) => void | Promise<void>;
-  /**
-   * Called after each page has been rendered and written to the output directory.
-   * @param page An object containing the path to the output file and its HTML content.
-   */
-  afterPage?: (page: { path: string; html: string }) => void | Promise<void>;
-  /**
-   * Called after the entire build process is complete.
-   * @param config The site configuration.
-   */
-  afterBuild?: (config: SiteConfig) => void | Promise<void>;
-}
-
-/**
- * The main orchestrator class for a Steno static site project.
- * It manages configuration loading, theme resolution, building markdown files to HTML, and dev server watching.
- *
- * @example
- * ```ts
- * import { Steno } from "@steno/steno";
- *
- * const steno = new Steno();
- * await steno.build();
- * ```
- */
-export class Steno {
-  private readonly config: SiteConfig;
-  private theme?: Theme;
-  private readonly themeLoadingPromise: Promise<void>;
-  private readonly autoBuildOnInit: boolean;
-  private plugins: StenoPlugin[] = [];
-  private readonly pluginsLoadingPromise: Promise<void>;
-
-  /**
-   * Creates a new Steno instance.
-   *
-   * @param configPath Path to the site config file. Defaults to `"content/.steno/config.yml"`.
-   * @param autoBuildOnInit If true, triggers a build immediately on instantiation unless running in dev mode.
-   * @param hooks Optional lifecycle hooks to tap into the build process.
-   */
-  constructor(
-    configPath: string = "content/.steno/config.yml",
-    autoBuildOnInit = true,
-    private hooks: StenoHooks = {},
-  ) {
-    this.config = loadConfig(configPath);
-    this.autoBuildOnInit = autoBuildOnInit;
-    this.themeLoadingPromise = this.loadTheme();
-    this.pluginsLoadingPromise = this.loadPlugins();
-    this.init();
-  }
-
-  /**
-   * Dynamically imports plugins declared in the site config and initializes them.
-   * @private
-   */
-  private async loadPlugins(): Promise<void> {
-    await this.themeLoadingPromise;
-    const sitePlugins = await loadPlugins(this.config);
-    this.plugins = [...(this.theme?.plugins ?? []), ...sitePlugins];
-  }
-
-  /**
-   * Dynamically imports the JSR/NPM theme package, or loads a local
-   * filesystem-based theme directory if a theme.yaml configuration file is present.
-   */
-  private async loadTheme(): Promise<void> {
-    const themeName = this.config.custom?.theme;
-    if (!themeName) return;
-
-    try {
-      const isLocalPath = themeName.startsWith(".") ||
-        themeName.startsWith("/") ||
-        themeName.startsWith("file://");
-
-      if (isLocalPath) {
-        const themeDir = themeName.startsWith("file://")
-          ? new URL(themeName).pathname
-          : (isAbsolute(themeName) ? themeName : join(Deno.cwd(), themeName));
-
-        // Check if theme.yaml or theme.yml exists inside the directory
-        let hasThemeYaml = false;
-        try {
-          Deno.statSync(join(themeDir, "theme.yaml"));
-          hasThemeYaml = true;
-        } catch {
-          try {
-            Deno.statSync(join(themeDir, "theme.yml"));
-            hasThemeYaml = true;
-          } catch {
-            // ignore
-          }
-        }
-
-        if (hasThemeYaml) {
-          // Load filesystem-based theme
-          this.theme = Theme.loadFromDirectory(
-            themeDir,
-            this.config.custom?.themeConfig,
-          );
-          return;
-        }
-
-        // Otherwise, import as a standard Deno file module
-        let resolvedPath = themeName.startsWith("file://")
-          ? themeName
-          : `file://${
-            isAbsolute(themeName) ? themeName : join(Deno.cwd(), themeName)
-          }`;
-
-        let stat: Deno.FileInfo | undefined;
-        try {
-          stat = Deno.statSync(new URL(resolvedPath));
-        } catch {
-          // Stat failed, try importing directly
-        }
-
-        if (stat?.isDirectory) {
-          let found = false;
-          for (const entry of ["mod.ts", "theme.ts", "index.ts"]) {
-            const testPath = `${resolvedPath.replace(/\/$/, "")}/${entry}`;
-            try {
-              Deno.statSync(new URL(testPath));
-              resolvedPath = testPath;
-              found = true;
-              break;
-            } catch {
-              // ignore
-            }
-          }
-          if (!found) {
-            console.error(
-              `Failed to load theme "${themeName}": Could not find mod.ts, theme.ts, or index.ts in theme directory "${themeName}"`,
-            );
-            return;
-          }
-        }
-        const themeModule = await import(resolvedPath);
-        const themeData = (themeModule.default || themeModule) as StenoTheme;
-        this.theme = new Theme(themeData, this.config.custom?.themeConfig);
-      } else {
-        // Dynamic import of remote JSR packages, npm packages, or URLs
-        const themeModule = await import(themeName);
-        const themeData = (themeModule.default || themeModule) as StenoTheme;
-        this.theme = new Theme(themeData, this.config.custom?.themeConfig);
-      }
-    } catch (error) {
-      console.error(`Failed to load theme "${themeName}":`, error);
-    }
-  }
-
-  /**
-   * Compiles the Markdown files into the final HTML output directory.
-   *
-   * @returns A promise that resolves when the build is complete.
-   */
-  public async build(): Promise<void> {
-    await this.themeLoadingPromise;
-    await this.pluginsLoadingPromise;
-
-    for (const plugin of this.plugins) {
-      await plugin.beforeBuild?.(this.config);
-    }
-    await this.hooks.beforeBuild?.(this.config);
-
-    const contentDir = this.config.contentDir || "content";
-    const outputDir = this.config.output || "dist";
-
-    ensureDirSync(outputDir);
-
-    // Pass 1: build collections
-    const collections: CollectionMap = await buildCollections(
-      contentDir,
-      this.config,
-      this.plugins,
-    );
-
-    // Pass 2: render pages
-    const processDirectory = async (currentDir: string, relPath = "") => {
-      for (const entry of Deno.readDirSync(currentDir)) {
-        const fullPath = join(currentDir, entry.name);
-        const entryRelPath = relPath ? join(relPath, entry.name) : entry.name;
-
-        if (entry.isDirectory) {
-          if (entry.name !== ".steno") {
-            await processDirectory(fullPath, entryRelPath);
-          }
-        } else if (entry.isFile && entry.name.endsWith(".md")) {
-          const fileContents = Deno.readTextFileSync(fullPath);
-
-          const { frontmatter, body } = parseFrontmatter(
-            fileContents,
-            fullPath,
-          );
-
-          let tokens = marked.lexer(body);
-          tokens = await runAstTransforms(tokens, this.plugins);
-          let htmlContent = marked.parser(tokens);
-          htmlContent = await runHtmlTransforms(htmlContent, this.plugins);
-
-          let outputFilePath = join(
-            outputDir,
-            entryRelPath.replace(/\.md$/, ".html"),
-          );
-          if (this.config.custom?.shortUrls) {
-            if (entryRelPath !== "index.md") {
-              const cleanRelPath = entryRelPath.replace(/\.md$/, "");
-              outputFilePath = join(outputDir, cleanRelPath);
-              ensureDirSync(outputFilePath);
-              outputFilePath = join(outputFilePath, "index.html");
-            } else {
-              outputFilePath = join(outputDir, "index.html");
-            }
-          } else {
-            ensureDirSync(dirname(outputFilePath));
-          }
-
-          const layoutName = typeof frontmatter.layout === "string"
-            ? frontmatter.layout
-            : "layout";
-
-          const renderedContent = this.theme
-            ? this.theme.renderLayout(layoutName, htmlContent, {
-              site: { ...this.config },
-              theme: {
-                name: this.theme.name,
-                version: this.theme.version,
-                ...this.theme.config,
-              },
-              collections, // ← injected here
-              title: frontmatter.title || this.config.title,
-              ...frontmatter,
-            })
-            : htmlContent;
-
-          Deno.writeTextFileSync(outputFilePath, renderedContent);
-
-          await this.hooks.afterPage?.({
-            path: outputFilePath,
-            html: renderedContent,
-          });
-          for (const plugin of this.plugins) {
-            await plugin.afterPage?.({
-              path: outputFilePath,
-              html: renderedContent,
-            });
-          }
-        }
-      }
-    };
-
-    await processDirectory(contentDir);
-
-    if (this.theme) {
-      await this.theme.copyAssets(outputDir);
-    }
-
-    for (const plugin of this.plugins) {
-      await plugin.afterBuild?.(this.config);
-    }
-
-    await this.hooks.afterBuild?.(this.config);
-
-    console.log("Build complete.");
-  }
-
-  /**
-   * Starts a development server with automatic file watching and rebuilding on change.
-   *
-   * @returns A promise that resolves when the dev server starts.
-   */
-  public async dev(): Promise<void> {
-    const contentDir = this.config.contentDir || "content";
-    const outputDir = this.config.output || "dist";
-    await startDevServer(outputDir, () => this.build(), contentDir);
-  }
-
-  /**
-   * Initializes the application.
-   */
-  private async init() {
-    if (this.autoBuildOnInit && !Deno.args.includes("dev")) {
-      await this.build();
-    }
-  }
-}
 
 if (import.meta.main) {
   try {
-    const options = parseCliArgs(Deno.args);
-
-    if (options.command === "help") {
-      printHelp();
-    } else if (options.command === "dev") {
-      const steno = new Steno(options.configPath, false);
-      await steno.dev();
-    } else {
-      const steno = new Steno(options.configPath, false);
-      await steno.build();
-    }
+    await runStenoCli(Deno.args);
   } catch (error) {
     console.error((error as Error).message);
     printHelp();
