+++ src/utils/cli_test.ts
