+++ src/utils/scribe.ts
