+++ test.ts
@@ -1,12 +1,13 @@
-import { registerCliTests } from "./src/cli_test.ts";
-import { registerConfigTests } from "./src/config_test.ts";
-import { registerFileUtilsTests } from "./src/fileUtils_test.ts";
-import { registerFrontmatterTests } from "./src/frontmatter_test.ts";
-import { registerScribeTests } from "./src/scribe_test.ts";
+import { registerCliTests } from "./src/utils/cli_test.ts";
+import { registerConfigTests } from "./src/core/config_test.ts";
+import { registerFileUtilsTests } from "./src/utils/fileUtils_test.ts";
+import { registerFrontmatterTests } from "./src/utils/frontmatter_test.ts";
+import { registerScribeTests } from "./src/utils/scribe_test.ts";
 import { registerThemeTests } from "./src/theme/theme_test.ts";
-import { registerBuildTests } from "./src/build_test.ts";
-import { registerPluginTests } from "./src/plugins_test.ts";
-import { registerCollectionTests } from "./src/collections_test.ts";
+import { registerServerTests } from "./src/utils/server_test.ts";
+import { registerBuildTests } from "./src/core/build_test.ts";
+import { registerPluginTests } from "./src/plugins/plugins_test.ts";
+import { registerCollectionTests } from "./src/core/collections_test.ts";
 
 registerCliTests();
 registerConfigTests();
@@ -14,6 +15,7 @@ registerFileUtilsTests();
 registerFrontmatterTests();
 registerScribeTests();
 registerThemeTests();
+registerServerTests();
 registerBuildTests();
 registerPluginTests();
 registerCollectionTests();
